Add NatSpec Documentation To The Safe Launchpad (#376)
Related to #308

This PR adds NatSpec documentation to the `SafeSignerLaunchpad` contract
for all methods. This is done in preparation for an auditing these
contracts and as a precursor to getting 100% test coverage.

In writing this PR, I found some issues with the current launchpad
implementation that need to be addressed pre-audit:
- We allow accounts that haven't been properly initialized, which can be
a security risk. We should do checks similar to the Safe `setup` method
to prevent double initialization and the validation should revert if the
account is not initialized.
- We are using revert messages instead of errors.

Additionally, I would like to suggest that we change the implementation
to use a **single** setup instead of doing a pre- and post- setup
delegatecall. This would greatly simplify the implementation and
security assumptions of the launchpad. The downside is that the setup
cannot break 4337 rules, but we do not have a use-case for this, so I
wouldn't implement it.