Moving `utils` from `test -> src` (#375)
Part of #351

> The code required to make the user story demos work (notably,
extracting the public key from a WebAuthn response and encoding the
WebAuthn signer signature) should move to the src/ tree (out of the
test/ tree)

This PR moves the files in the `test/utils` folder to `src/utils`.
Corresponding changes to the test files also are made. This is a
pre-requisite for the NPM module publishing.