Support Precompile With Fallback For Signers (#379)
This PR changes the internal workings of the signers to accept a uint192
instead of an address as the `verifier` (well, now `verifiers`)
parameter. This allows a signer to encode BOTH a precompile address as
well as a Solidity implementation fallback for the P-256 curve.

Additionally, this required changing the interface of the
`ICustomECDSASignerFactory`. It is now very much tied to the P-256
implementation that we have, and as such I renamed things to just
`ISafeSignerFactory` and `SafeSignerLaunchpad`, since it is not a
general ECDSA solution, but one that is particular to the P-256
implementation that we have and in particular supporting P-256
precompile with Solidity fallback.

The PR changes **a lot** of files, but this is mostly small
modifications to tests. Unfortunately, I don't really see a way around
this without pushing a PR with a broken test suite which I did not want
to do.

Overall, I think this is definitely a nice feature to have - as it adds
a small overhead on top of the regular verification but allows accounts
to seamlessly start using precompile on networks that don't have it
available yet.