Use Account Storage for Shared WebAuthn Signer (#412)
Fixes #408

This PR reworks the WebAuthn singleton signer to use the account storage
instead of its own storage for signer configuration (x, y, and verifiers
configuration values). This allows the shared signer to work **without**
a staked factory 🎉!

We also rename the methods a bit to be more consistent with our signer
factory implementation, and adjust the E2E test accordingly (noting that
a staked factory is no longer needed).

This also required some small adjustments to the
`4337-passkeys-singleton-signer` example for the new code (but overall
simplification - no need for a staked factory anymore).