Add Additional ERC-4337 Related Passkey Tests (#417)
Follow up to #415

Now that we officially tagged the launchpad as "experimental", the
module was missing unit and E2E tests demonstrating the "officially"
supported way to use passkeys over ERC-4337. This PR adds both a unit
test that a user operation with a passkey owner can be executed over the
ERC-4337 reference entrypoint, as well as a local bundler test to ensure
that no bundler rules are violated by the passkey signers.