Add Typescript dependency to passkey package (#421)
This PR adds missing `typescript` dependency to the `safe-passkey`
package which was causing build errors locally (maybe there is a global
Typescript installation on CI and on @mmv08's machine). Also, there was
some linting errors that weren't caught.

Also, PNPM reordered `package.json` for the passkey package, and I
manually did it for the other packages.