Local Bundler Test with New Shared Signer Flow (#423)
This PR demonstrates that the launchpad is, in fact, kind of useless.

Specifically, a new local bundler end-to-end test is added where the
shared signer is used to create a Safe that ends up with the same
configuration as with the signer launchpad. It uses the shared signer as
the initial owner which is used for signing **only** the first user
operation. During execution of the first user operation, the Safe's
owner is swapped for a freshly deployed `SafeWebAuthnSignerProxy`
instead of the shared signer. This means that we end up with a Safe with
a "regular" passkey owner without any of the downsides (complicated
launchpad code, rigid ownership structure, only one passkey owner
allowed, etc.).

If this PR is approved, I will **remove** the `SafeSignerLaunchpad` in
follow up PRs as well as update the passkey example to use the same
strategy so it ends up with a standard passkey owner.