Shared Signer Mapping
This PR changes the behaviour of the shared signer to store its
configuration as if it were behind a mapping where the key is the shared
signer address. This allows multiple shared signer instances to be
configured for a single account (in case you ever need to deploy a 2/2
Safe owned by passkeys with 4337).

I created an "accessor" contract to prove that the storage location is
indeed the equivalent to a `mapping(address self => Signer)`.