Fix `runOp` Script (#435)
I was running the 4337 `runOp.ts` script this morning and noticed it was
subtly broken. This PR fixes it.

I also refactored the scripts a bit - specifically we no longer `--fix`
on `lint:ts`. This was because it would cause CI to silently pass in
cases where the linter could autofix the code (when we want it to fail
instead).