Add E2E Test For Multiple Passkey Owners With 4337 (#432)
Fixes #372

This PR adds a new E2E test that demonstrates the use of a Safe being
deployed over 4337 with multiple passkey owners. This means that we can
have arbitrary ownership structures with passkeys and 4337 :tada:.

There is an important caveat however, the tracing mechanism used by the
bundler is timing out when verifying multiple passkey signatures, so
this E2E test is only half complete. At least we verify that the
signature verification logic would work, and eventually when precompiles
become part of Geth, we will be able to switch the E2E test to use that
instead which should greatly speed up tracing.

I still think it is worth merging this PR as is, as it also checks that
multiple shared signer instances can be used in parallel, a new feature
added in #429.