Deprecate `SafeSignerLaunchpad` (#434)
This PR removes the `SafeSignerLaunchpad` as we can provide a better
developer experience with just the `SafeWebAuthnSharedSigner` contract
when working with ERC-4337.

This PR also takes the opportunity to move the
`4337-passkeys-singleton-signer` example to `4337-passkeys`. I noticed
that the former was not being built in CI, so there were a couple of
tweaks necessary to get it building (related to `package.json` changes
required to make it build with PNPM).

While it looks like there are a lot of changes in this PR, essentially
all it does is:
1. Remove the `SafeSignerLaunchpad` contract (`git rm
modules/passkey/contracts/4337/experimental/SafeSignerLaunchpad.sol`) as
well as the related deployment script and tests
2. Untag the 4337 `SafeWebAuthnSharedSigner` contract as experimental
(`git mv
modules/passkey/contracts/4337/experimental/SafeWebAuthnSharedSigner.sol
modules/passkey/contracts/4337/`)
3. Remove the old `4337-passkeys` example application (`git rm -r
examples/4337-passkeys`)
4. Move the `4337-passkeys-singleton-signer` example application to
replace the `4337-passkeys` one as the canonical example app (`git mv
examples/4337-passkeys-singleton-signer examples/4337-passkeys`)

No logic changes were made besides moving code around.