4337 Benchmark: Alchemy with EntryPoint 0.7.0 (#466)
This PR partially #316:
1. It upgrades Alchemy code to use the v0.7.0 of the entrypoint,
including all the alchemy-related packages
2. I exported all the alchemy-related utils function from
`src/utils/userOps.ts` to `alchemy/utils.ts`
3. I also added a special exception for the case where the user
operations do not involve account deployment where if the account
doesn't have enough stake in the entrypoint from the previous deployment
user operation, it will execute a call to `depositTo` in the entry point
and inflate the benchmark numbers. Now, a special case is added with the
entrypoint deposit before the user operation submission for all user
operations that do not involve a deployment.