Fix Flaky WebAuthn Shim Test
Fixes #369

This PR bumps the version of `@simplewebauthn/server` to include a fix
to which was causing flacky tests for our WebAuthn shim. It turns out
that the library was not correctly dealing with leading zeros in the
signature `r` or `s` value, which would cause the test to periodically
fail (since the signature was random in each run - as both the private
kay and the challenge were randomly generated).